requirements.txt
python-docx==1.1.2
langchain==0.3.19
openai==1.65.2
google-generativeai==0.8.4
deepseek-ai==0.0.1
python-dotenv==1.0.1
streamlit==1.42.2 